Explicitly cast double to uint32, to suppress warning

+++ src/arping.c
@@ -1227,7 +1227,10 @@ wait_time(double deadline, uint32_t packetwait)
         if (max_wait > packetwait / 1000000.0) {
                 return packetwait;
         }
-        return max_wait * 1000000;
+
+        // TODO: check for overflows. Though what's the worst that
+        // could happen?
+        return (uint32_t)(max_wait * 1000000);
 }
 
 /**
